feat: composable use grid table
